requirements.txt
requests==2.23.0
environs==8.0.0
python-telegram-bot==12.7
redis==3.5.3